% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_full_new.R
\name{col_scale}
\alias{col_scale}
\title{Center and/or normalize each column of a matrix}
\usage{
col_scale(X, center = T, normalize = F)
}
\arguments{
\item{X}{a numeric matrix with NAs or "Incomplete" matrix object (see softImpute package)}

\item{center}{center each column of \code{X} if \code{center == TRUE}. 
The default value is \code{TRUE}.}

\item{normalize}{normalize each column of \code{X} such that its sample variance is 1 if \code{normalize == TRUE}. 
The default value is \code{False}.}
}
\value{
a centered and/or normalized matrix of the same dimension as \eqn{X}.
}
\description{
Center and/or normalize each column of a matrix
}
