\name{primePCA-package}
\alias{primePCA-package}
\docType{package}
\title{
\packageTitle{primePCA}
}
\description{
\packageDescription{primePCA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{primePCA}
\packageIndices{primePCA}
}

\author{
\packageAuthor{primePCA}

Maintainer: \packageMaintainer{primePCA}
}
\references{
Zhu, Z., Wang, T. and Samworth, R. J. (2019) High dimensional principal component analysis with heterogeneous missingness. <arXiv:1906.12125>
}
\keyword{package}
