% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_full_new.R
\name{sin_theta_distance}
\alias{sin_theta_distance}
\title{Frobenius norm sin theta distance between two column spaces}
\usage{
sin_theta_distance(V1, V2)
}
\arguments{
\item{V1}{a matrix with orthonormal columns}

\item{V2}{a matrix of the same dimension as V1 with orthonormal columns}
}
\value{
the Frobenius norm sin theta distance between two V1 and V2
}
\description{
Frobenius norm sin theta distance between two column spaces
}
